requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-filter==2.4.0
django-heroku==0.0.0
djangorestframework==3.12.4
Faker==8.1.2
gunicorn==20.0.4
Markdown==3.3.4
mysql-connector-python==8.0.24
mysqlclient==2.0.3
protobuf==3.15.8
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
